api/requirements.txt
openai
langchain
flask
flask-cors
langchain-community
langchain-deepseek-official
transformers
pypdfium2